+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 * Add inclusive/exclusive options to clj-toolbox.colls/take-range
+* Convert some extraneous macros in clj-toolbox.test-utils to be a function instead
 
 ## [2.0.0] — 2025-04-30
 * Remove deprecated functions from clj-toolbox.files